requirements.txt
pandas==1.1.1
PyMySQL==0.9.3
SQLAlchemy==1.3.19
numpy==1.19.1
Flask_Cors==3.0.8
bokeh==2.3.1
Flask==1.0.2
PyJWT==2.1.0
